third_party/WebKit/LayoutTests/fast/css/invalidation/explicit-inheritance-propagation-expected.txt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS getComputedStyle(inner).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
